fix: dynamic word route segment replaced with query params below /search path
so that we can avoid guarding these pages
